+++ IOT/src/Makefile
@@ -1,6 +1,6 @@
 CC = gcc
-LIBS = -lbluetooth -lpthread -lcsv -lproperties
-INC = -Lcsv_writer -Lproperties_pack
+LIBS = -lbluetooth -lpthread -lcsv -lproperties -lblue
+INC = -Lcsv_writer -Lproperties_pack -Lbluetooth_pack
 TARGET = ../airbeat.exe 
 OBJECT = main_service.o 
 all : $(TARGET)
