+++ IOT/src/header.h
@@ -1,4 +1,5 @@
 #include<unistd.h>
+#include<stdint.h>
 #include<stdio.h>
 #include<string.h>
 #include<stdlib.h>
@@ -8,4 +9,3 @@
 #include<sys/socket.h>
 #include<sys/types.h>
 #include<sys/stat.h>
-
